requirements.txt
# Core dependencies for Qwen Vision-Language Realtime Service
# Python 3.10+

# Web framework and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# WebRTC and media processing
aiortc==1.6.0
aiohttp==3.9.1
av==11.0.0

# Computer vision and image processing
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# ML and model serving
# vLLM ROCm - use system install or specific wheel
# vllm==0.9.2  # Uncomment if pip-installable for ROCm
torch>=2.1.0
transformers==4.37.0
accelerate==0.26.0
sentencepiece==0.1.99
protobuf==4.25.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Monitoring and logging
psutil==5.9.7
prometheus-client==0.19.0

# Optional TTS (lightweight)
# piper-tts==1.2.0  # Uncomment if TTS needed

# Development tools
black==24.1.1
pytest==7.4.4
pytest-asyncio==0.23.3

